Refactor deaths radar chart into feature-based architecture
Reorganized deaths radar chart from scattered data-tracking structure into cohesive feature-based organization under analysis/deaths-radar/, extracting radar-specific calculations into dedicated logic directory.

Maintains feature isolation pattern established in previous migrations (tier-stats, tier-trends, game-runs), improving discoverability and preparing structure for future radar enhancements.
