+++ src/features/data-tracking/utils/chart-data.ts
@@ -1,15 +1,12 @@
 import { format, startOfDay, startOfWeek, startOfMonth, startOfYear } from 'date-fns'
 import { ParsedGameRun } from '../types/game-run.types'
-import { getFieldValue } from './field-utils'
 import { RunTypeFilter, filterRunsByType } from './run-type-filter'
-import { 
-  ChartDataPoint, 
-  TimePeriod, 
-  TimePeriodConfig, 
+import {
+  ChartDataPoint,
+  TimePeriod,
+  TimePeriodConfig,
   TIME_PERIOD_CONFIGS,
-  KilledByData,
-  TierKilledByData,
-  TierStatsData 
+  TierStatsData
 } from './chart-types'
 import {
   prepareCoinsPerRunData,
@@ -38,78 +35,8 @@ export {
   prepareYearlyData
 } from './date-aggregation'
 
-export function prepareKilledByData(runs: ParsedGameRun[], runTypeFilter: RunTypeFilter = 'all'): TierKilledByData[] {
-  // Filter runs by type first
-  const filteredRuns = filterRunsByType(runs, runTypeFilter);
-
-  // Group runs by tier
-  const tierGroups = new Map<number, ParsedGameRun[]>()
-
-  filteredRuns.forEach(run => {
-    const killedBy = getFieldValue<string>(run, 'killedBy');
-    if (run.tier && killedBy) {
-      if (!tierGroups.has(run.tier)) {
-        tierGroups.set(run.tier, [])
-      }
-      tierGroups.get(run.tier)!.push(run)
-    }
-  })
-
-  // Process each tier's killed-by data
-  const tierData: TierKilledByData[] = []
-
-  tierGroups.forEach((tierRuns, tier) => {
-    // Count deaths by type
-    const deathCounts = new Map<string, number>()
-
-    tierRuns.forEach(run => {
-      const killedBy = getFieldValue<string>(run, 'killedBy') || 'Unknown'
-      deathCounts.set(killedBy, (deathCounts.get(killedBy) || 0) + 1)
-    })
-
-    const totalDeaths = tierRuns.length
-
-    // Convert to percentage-based data for radar chart
-    const killedByStats: KilledByData[] = Array.from(deathCounts.entries())
-      .map(([killedBy, count]) => ({
-        killedBy,
-        count,
-        percentage: (count / totalDeaths) * 100
-      }))
-      .sort((a, b) => b.count - a.count) // Sort by count descending
-
-    tierData.push({
-      tier,
-      killedByStats,
-      totalDeaths
-    })
-  })
-
-  return tierData.sort((a, b) => a.tier - b.tier)
-}
-
-// Prepare radar chart data by combining all tiers' top death causes
-export function prepareRadarChartData(tierData: TierKilledByData[]): any[] {
-  // Get all unique death causes across all tiers
-  const allDeathCauses = new Set<string>()
-  tierData.forEach(tier => {
-    tier.killedByStats.forEach(stat => {
-      allDeathCauses.add(stat.killedBy)
-    })
-  })
-
-  // Create radar chart data points
-  return Array.from(allDeathCauses).map(deathCause => {
-    const dataPoint: any = { killedBy: deathCause }
-
-    tierData.forEach(tier => {
-      const stat = tier.killedByStats.find(s => s.killedBy === deathCause)
-      dataPoint[`tier${tier.tier}`] = stat ? stat.percentage : 0
-    })
-
-    return dataPoint
-  })
-}
+// Re-export radar chart functions from their new feature location
+export { prepareKilledByData, prepareRadarChartData } from '../../analysis/deaths-radar/logic/radar-calculations'
 
 export function prepareTierStatsData(runs: ParsedGameRun[], runTypeFilter: RunTypeFilter = 'all'): TierStatsData[] {
   // Filter runs by type first
